Extract BuildEmail form into separate component
